Extract and export volume modes as separate module
